+++ browser/fetcher.js
@@ -2,6 +2,7 @@
 
 var xhr = require('./xhr');
 var state = require('./state');
+var router = require('./router');
 var emitter = require('./emitter');
 var deferral = require('./deferral');
 var interceptor = require('./interceptor');
@@ -17,7 +18,7 @@ function negotiate (route, context) {
   var qs = e(parts.search);
   var p = qs ? '&' : '?';
   var demands = ['json'].concat(deferral.needs(route.action));
-  if (context.hijacker) {
+  if (context.hijacker && context.hijacker !== route.action) {
     demands.push('hijacker=' + context.hijacker);
   }
   return parts.pathname + qs + p + demands.join('&');
@@ -55,7 +56,7 @@ function fetcher (route, context, done) {
     }
   }
 
-  function notify (err, data) {
+  function notify (err, data, res) {
     if (err) {
       global.DEBUG && global.DEBUG('[fetcher] failed for %s', route.url);
       if (err.message === 'aborted') {
@@ -67,7 +68,7 @@ function fetcher (route, context, done) {
       global.DEBUG && global.DEBUG('[fetcher] succeeded for %s', route.url);
       if (data && data.version) {
         state.version = data.version; // sync version expectation with server-side
-        componentCache.set(route.action, data);
+        componentCache.set(router(res.url).parts.query.hijacker || route.action, data);
       }
       emitter.emit('fetch.done', route, context, data);
     }