+++ browser/once.js
@@ -1,6 +1,6 @@
 'use strict';
 
-module.exports = function (fn) {
+module.exports = function disposable (fn) {
   var used;
   return function once () {
     if (used) { return; } used = true;